refactor: move TemplateWrapper into Core
