+++ resources/views/layout/base.php
@@ -1,6 +1,6 @@
 <?php
 
-use Vibrato\ThemeTemplateWrapper;
+use Vibrato\Core\TemplateWrapper;
 use Vibrato\ThemeSetup;
 ?>
 
@@ -19,14 +19,14 @@ use Vibrato\ThemeSetup;
             <div class="max-w-7xl mx-auto py-6 sm:px-6 lg:px-8">
                 <!-- Replace with your content -->
                 <div class="px-4 py-6 sm:px-0">
-                    <?php include ThemeTemplateWrapper::template_path(); ?>
+                    <?php include TemplateWrapper::template_path(); ?>
                 </div>
                 <!-- /End replace -->
             </div>
         </main><!-- /.main -->
         <?php if (ThemeSetup::display_sidebar()) : ?>
             <aside class="sidebar">
-                <?php include ThemeTemplateWrapper::sidebar_path(); ?>
+                <?php include TemplateWrapper::sidebar_path(); ?>
             </aside><!-- /.sidebar -->
         <?php endif; ?>
     </div><!-- /.wrap -->
